requirements.txt
scikit-learn==1.6.1
opencv-python-headless
mediapipe==0.10.13
fastapi
uvicorn
numpy
pyspellchecker
